bugzilla#222: refactor wiki links and fix WikiPage.get_links
Now Reference nodes only have one extra key when they are wiki links:
"wiki_name".  It is False if the link is not resolved, or the path from the
source page to the destination page.

If you need the original name given in backquotes, use the "name" key provided
by docutils.