giving an icon to views: allow to decorate subtabs and context menu
You no longer have to add your CSS rule in Aruni or in your skin to give an
icon to new views.

It also removes the link between the label and the tab id: preventing French or
Chinese labels to produce bogus CSS ids, whilst still inheriting the base icon.