bug#1019: configure non-root modules at init time
With less and less projects overriding the root class, and as modularity
goes (e.g. wiki), we need a way to create a functional instance in a
single command line.

It will ease up documentation by not requiring people to edit the config
by hand.

And usine will appreciate it too.