Add FileField, and use it for users
Start moving resource.handler to the (new) schema.

The new module fields.py will, in some unknown future, replace the
'class_schema' stuff. For now it just contains the 'FileField' class,
which allows a resource to define files attached to it.

For instance we use it for the user's avatar now, the new schema looks
like this:

  class User(DBResource):
      avatar = FileField(title=MSG(u'Avatar'))

And a user in the database looks like this:

  0.metadata
  0.avatar

Now users are not folders.